requirements.txt
python-telegram-bot==20.3
openai==1.3.0
PyYAML==6.0
APScheduler==3.10.4
requests==2.31.0